+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## 0.24.0
 
 ### Enhancements
+- Designs can now be copied from one project to another
+  - Download the JSON version of the design (or designs) from origin project
+  - Use the Import Design from JSON to import the designs
+  - Import won't overright existing designs/variables/domains if  design/variable/domain of the same name already exists
 - **Gem Changes**
   - Updated to rails 4.1.0.rc2
   - Updated to contour 2.5.0.beta1